# wwwroot/swagger/v1/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Api Comercial",
    "description": "API para la gestión de documentos comerciales",
    "contact": {
      "name": "Soporte API",
      "email": ""
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Vehiculos": {
      "get": {
        "tags": [
          "Vehiculos"
        ],
        "summary": "Obtener todos los vehículos.",
        "operationId": "Vehiculos_GetVehiculos",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseVehiculo"
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vehiculos"
        ],
        "summary": "Insertar un nuevo vehículo.",
        "operationId": "Vehiculos_InsertVehiculo",
        "requestBody": {
          "x-name": "parametros",
          "description": "Datos del vehículo a insertar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertarVehiculoRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Vehiculos/{idChasis}": {
      "get": {
        "tags": [
          "Vehiculos"
        ],
        "summary": "Obtener un vehículo por ID de chasis.",
        "operationId": "Vehiculos_GetVehiculoPorId",
        "parameters": [
          {
            "name": "idChasis",
            "in": "path",
            "required": true,
            "description": "Identificador del chasis del vehículo",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseVehiculo"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Vehiculos"
        ],
        "summary": "Actualizar los datos de un vehículo en su totalidad.",
        "operationId": "Vehiculos_UpdateVehiculo",
        "parameters": [
          {
            "name": "idChasis",
            "in": "path",
            "required": true,
            "description": "Identificador del chasis del vehículo",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "parametros",
          "description": "Datos del vehículo a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestVehiculo"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Vehiculos"
        ],
        "summary": "Eliminar un vehículo por ID de chasis.",
        "operationId": "Vehiculos_DeleteVehiculo",
        "parameters": [
          {
            "name": "idChasis",
            "in": "path",
            "required": true,
            "description": "Identificador del chasis del vehículo a eliminar",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Vehiculos/estados": {
      "get": {
        "tags": [
          "Vehiculos"
        ],
        "summary": "Endopoint para consultar estados de vehiculos",
        "operationId": "Vehiculos_GetEstados",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EstadosResponse"
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vehiculos"
        ],
        "summary": "Endpoint encargado de insertar los estados para los vehiculos",
        "operationId": "Vehiculos_InsertEstados",
        "requestBody": {
          "x-name": "parametros",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstadosRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/estado/{estado}": {
      "get": {
        "tags": [
          "Vehiculos"
        ],
        "summary": "Endpoint para obtener el conteo de vehículos por estado.",
        "operationId": "Vehiculos_GetCountByEstado",
        "parameters": [
          {
            "name": "estado",
            "in": "path",
            "required": true,
            "description": "El estado de los vehículos.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "El número de vehículos que tienen el estado especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Vehiculos/total-vehiculos": {
      "get": {
        "tags": [
          "Vehiculos"
        ],
        "summary": "Devuelte el total de vehiculos, excluyendo los vendidos.",
        "operationId": "Vehiculos_GetTotalVehiculos",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Vehiculos/detalleVenta": {
      "post": {
        "tags": [
          "Vehiculos"
        ],
        "summary": "Endpoint que inserta el detalle de la venta realizada",
        "operationId": "Vehiculos_InsertarDetalleVentas",
        "requestBody": {
          "x-name": "ventaDetalleRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetalleVentaRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Clase que muestra los errores",
        "additionalProperties": false,
        "properties": {
          "errorType": {
            "$ref": "#/components/schemas/ErrorType"
          },
          "errorDescripcion": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ResponseVehiculo": {
        "type": "object",
        "description": "Clase de respuesta de vehiculos",
        "additionalProperties": false,
        "properties": {
          "idChasis": {
            "type": "string",
            "description": "Numero de chasis del vehiculo",
            "nullable": true
          },
          "idMarca": {
            "type": "integer",
            "description": "Codigo de la marca del vehiculo",
            "format": "int32",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "description": "Marca del vechiculo",
            "nullable": true
          },
          "idModelo": {
            "type": "integer",
            "description": "Codigo del modelo",
            "format": "int32",
            "nullable": true
          },
          "modelo": {
            "type": "string",
            "description": "Modelo del vehiculo",
            "nullable": true
          },
          "tipoCar": {
            "type": "string",
            "description": "Tipo del vehiculo",
            "nullable": true
          },
          "anoFabricacion": {
            "type": "string",
            "description": "Año de fabricación"
          },
          "color": {
            "type": "string",
            "description": "Color del vehiculo",
            "nullable": true
          },
          "usado": {
            "type": "string",
            "nullable": true
          },
          "chapa": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "string",
            "nullable": true
          },
          "despacho": {
            "type": "string",
            "nullable": true
          },
          "costo": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "EstadosResponse": {
        "type": "object",
        "description": "Muestra los datos de estados",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id del estado",
            "format": "int32"
          },
          "descripcion": {
            "type": "string",
            "description": "Descripcion del estado"
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Ingrese: Bearer {su token JWT}",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}